# docs/swagger.yaml
basePath: /
definitions:
  server.PaginatedTreeResponse:
    properties:
      children:
        items:
          $ref: '#/definitions/server.TreeChildNode'
        type: array
      kind:
        type: string
      modifiedAt:
        type: string
      name:
        type: string
      pagination:
        $ref: '#/definitions/server.PaginationMetadata'
      path:
        type: string
      size:
        type: integer
      summary:
        $ref: '#/definitions/server.TreeStatisticsSummary'
    type: object
  server.PaginationMetadata:
    properties:
      hasMore:
        type: boolean
      limit:
        type: integer
      nextPageUrl:
        type: string
      offset:
        type: integer
      prevPageUrl:
        type: string
      total:
        type: integer
    type: object
  server.TreeChildNode:
    properties:
      kind:
        type: string
      link:
        type: string
      modifiedAt:
        type: string
      name:
        type: string
      path:
        type: string
      size:
        type: integer
    type: object
  server.TreeStatisticsSummary:
    properties:
      directoryCount:
        type: integer
      fileCount:
        type: integer
      largestChildren:
        items:
          $ref: '#/definitions/server.TreeChildNode'
        type: array
      totalChildren:
        type: integer
      totalSize:
        type: integer
    type: object
info:
  contact:
    name: API Support
    url: https://github.com/prismon/mcp-space-browser
  description: |-
    Disk space indexing agent with REST API and MCP tools for exploring filesystem utilization

    This API provides endpoints for indexing filesystems and retrieving hierarchical disk space data.
    The server also exposes MCP (Model Context Protocol) tools at /mcp for advanced disk space analysis.
  license:
    name: MIT
  title: MCP Space Browser API
  version: 0.1.0
paths:
  /api/index:
    get:
      consumes:
      - application/json
      description: Triggers asynchronous indexing of the specified filesystem path.
        The indexing process crawls the directory tree and stores metadata in the
        database.
      parameters:
      - description: Filesystem path to index
        example: '"/home/user/Documents"'
        in: query
        name: path
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: path required
          schema:
            type: string
      summary: Index a filesystem path
      tags:
      - Index
  /api/tree:
    get:
      consumes:
      - application/json
      description: Returns a paginated directory listing with statistics and summary
        information
      parameters:
      - description: Filesystem path to analyze (defaults to current directory)
        example: '"/home/user/Documents"'
        in: query
        name: path
        type: string
      - description: 'Maximum number of children to return per page (default: 100,
          max: 1000)'
        example: 100
        in: query
        name: limit
        type: integer
      - description: 'Pagination offset (default: 0)'
        example: 0
        in: query
        name: offset
        type: integer
      - description: 'Sort children by: size, name, mtime (default: size)'
        example: '"size"'
        in: query
        name: sortBy
        type: string
      - description: 'Sort order: asc or desc (default: desc)'
        example: '"desc"'
        in: query
        name: order
        type: string
      - description: 'Include children in response (default: true)'
        example: true
        in: query
        name: includeChildren
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.PaginatedTreeResponse'
        "400":
          description: invalid path
          schema:
            type: string
        "500":
          description: failed to build tree
          schema:
            type: string
      summary: Get hierarchical directory tree with pagination
      tags:
      - Tree
schemes:
- http
- https
swagger: "2.0"
tags:
- description: Filesystem indexing operations
  name: Index
- description: Hierarchical disk space data retrieval
  name: Tree